@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,36 +28,85 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: any
+  font_awesome_flutter: any
+  flutter_svg: any
+  page_transition: any
+  intro_views_flutter: any
+  carousel_slider: ^4.1.1
+  # test: any
+
+  #Dialog
+  simple_animations: any
+  flare_flutter: any
+  top_snackbar_flutter: ^1.1.1
+  #Gird View card
+  flutter_layout_grid: any
+  flutter_staggered_grid_view: any
+  flutter_lints: any
+
+  #Firebase Dependencies
+  cloud_firestore: any
+  firebase: any
+  firebase_auth: any
+  firebase_core: any
+  firebase_storage: any
+  firebase_analytics: any
+  fluttertoast: any
+  unicons: any
+  timeago: any
+  intl: ^0.17.0
+
+  #State Management
+  flutter_bloc: any
+  image_picker: any
+  image_cropper: any
+
+  #Hive DB and SQLite DB
+  path: any
+  bloc: any
+  equatable: any
+  http: ^0.13.4 
+  sqflite: any
+  sqlite3: any
+
+  #Animation
+  shimmer_animation: any
+  animations: any
+  expandable: any
+  flutter_neumorphic: ^3.2.0
+
+  get: any
   google_sign_in: any
   flutter_icons: any
   percent_indicator: any
   cached_network_image: any
   flutter_spinkit: any
-  cloud_firestore: any
-  fluttertoast: any
-  animations: any
   image: any
   uuid: any
-  firebase_auth: any
-  firebase_core: any
-  image_picker: any
-  firebase_storage: any
   path_provider: any
   geolocator: any
   geocoding: any
-  timeago: any
   google_fonts: any
   provider: any
   modal_progress_hud: any
-  image_cropper: any
   shared_preferences: any
   data_connection_checker: any
+  flutter_slidable: any
+  flutter_typeahead: any
+  analyzer: 
+  translator: 
+  # google_translator: ^1.0.0
 
 dev_dependencies:
+  integration_test:
+    sdk: flutter
+  test: any
+
   flutter_test:
     sdk: flutter
+  mockito: any
+  build_runner: any
 
-  flutter_launcher_name: any
 
 flutter_launcher_name:
   name: "medverse_mobile_app"
@@ -74,7 +123,15 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
+    - assets/image/300_imagesrxnav/
+    - assets/icons/
+    - assets/flare/
+    - database/Glossary.sqlite
+    - database/drugbank.sqlite3
     - assets/images/
+    - assets/images/splash/Medverse.png
+    - assets/health_profile_icon/
+
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
