@@ -2,8 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:provider/provider.dart';
+import '/theme/palette.dart';
+import '/utils/app_text_theme.dart';
 import '/components/custom_image.dart';
-import '/view_models/auth/posts_view_model.dart';
+import '../../view_models/post/posts_view_model.dart';
 import '/widgets/indicators.dart';
 
 class ProfilePicture extends StatefulWidget {
@@ -26,7 +28,11 @@ class _ProfilePictureState extends State<ProfilePicture> {
         child: Scaffold(
           key: viewModel.scaffoldKey,
           appBar: AppBar(
-            title: Text('Add a profile picture'),
+            backgroundColor: Palette.mainBlueTheme,
+            title: Text(
+              'Thêm một ảnh đại diện',
+              style: MobileTextTheme().appBarStyle,
+            ),
             centerTitle: true,
           ),
           body: ListView(
@@ -38,12 +44,12 @@ class _ProfilePictureState extends State<ProfilePicture> {
                   width: MediaQuery.of(context).size.width,
                   height: MediaQuery.of(context).size.width - 30,
                   decoration: BoxDecoration(
-                    color: Colors.grey[300],
+                    color: Palette.grey300,
                     borderRadius: BorderRadius.all(
                       Radius.circular(3.0),
                     ),
                     border: Border.all(
-                      color: Theme.of(context).accentColor,
+                      color: Palette.mainBlueTheme,
                     ),
                   ),
                   child: viewModel.imgLink != null
@@ -56,10 +62,8 @@ class _ProfilePictureState extends State<ProfilePicture> {
                       : viewModel.mediaUrl == null
                           ? Center(
                               child: Text(
-                                'upload your profile picture',
-                                style: TextStyle(
-                                  color: Theme.of(context).accentColor,
-                                ),
+                                'Tải ảnh đại diện của bạn lên đây',
+                                style: MobileTextTheme().choosePictureRequired,
                               ),
                             )
                           : Image.file(
@@ -74,8 +78,8 @@ class _ProfilePictureState extends State<ProfilePicture> {
               Center(
                 child: ElevatedButton(
                   style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.all<Color>(
-                        Theme.of(context).accentColor),
+                    backgroundColor:
+                        MaterialStateProperty.all<Color>(Palette.mainBlueTheme),
                     shape: MaterialStateProperty.all<RoundedRectangleBorder>(
                       RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(5.0),
@@ -83,7 +87,10 @@ class _ProfilePictureState extends State<ProfilePicture> {
                     ),
                   ),
                   child: Center(
-                    child: Text('done'.toUpperCase()),
+                    child: Text(
+                      'Đã hoàn tất',
+                      style: AppTextTheme.oswaldTextStyle,
+                    ),
                   ),
                   onPressed: () => viewModel.uploadProfilePicture(context),
                 ),
@@ -111,7 +118,7 @@ class _ProfilePictureState extends State<ProfilePicture> {
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 10.0),
                 child: Text(
-                  'Select'.toUpperCase(),
+                  'Lựa chọn',
                   style: TextStyle(
                     fontWeight: FontWeight.bold,
                   ),
@@ -120,7 +127,10 @@ class _ProfilePictureState extends State<ProfilePicture> {
               Divider(),
               ListTile(
                 leading: Icon(Feather.camera),
-                title: Text('Camera'),
+                title: Text(
+                  'Chọn từ máy ảnh',
+                  style: MobileTextTheme().selectCamera,
+                ),
                 onTap: () {
                   Navigator.pop(context);
                   viewModel.pickImage(camera: true);
@@ -128,7 +138,10 @@ class _ProfilePictureState extends State<ProfilePicture> {
               ),
               ListTile(
                 leading: Icon(Feather.image),
-                title: Text('Gallery'),
+                title: Text(
+                  'Chọn từ thư viện',
+                  style: MobileTextTheme().selectGallery,
+                ),
                 onTap: () {
                   Navigator.pop(context);
                   viewModel.pickImage();
