@@ -41,7 +41,7 @@ class StreamBuilderWrapper extends StatelessWidget {
                     height: 60.0,
                     width: 100.0,
                     child: Center(
-                      child: Text('No Posts'),
+                      child: Text('Hiện chưa có bài viết nào!'),
                     ),
                   ),
                 )
