@@ -36,8 +36,8 @@ class CommentsStreamWrapper extends StatelessWidget {
               ? Container(
                   child: Center(
                     child: Padding(
-                      padding: const EdgeInsets.only(top: 20.0),
-                      child: Text('No comments'),
+                      padding: const EdgeInsets.only(top: 20.0, bottom: 20.0),
+                      child: Text('Hiện chưa có bình luận nào'),
                     ),
                   ),
                 )
