@@ -4,11 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:provider/provider.dart';
+import '/widgets/awesome_dialog.dart';
 import '/components/custom_image.dart';
 import '/models/user.dart';
 import '/utils/firebase.dart';
-import '/view_models/auth/posts_view_model.dart';
+import '/utils/validation.dart';
+import '/utils/app_text_theme.dart';
+import '../view_models/post/posts_view_model.dart';
 import '/widgets/indicators.dart';
+import '/theme/palette.dart';
 
 class CreatePost extends StatefulWidget {
   @override
@@ -16,6 +20,10 @@ class CreatePost extends StatefulWidget {
 }
 
 class _CreatePostState extends State<CreatePost> {
+  /// Editing Controllers
+  TextEditingController description = TextEditingController();
+  bool loading = false;
+
   @override
   Widget build(BuildContext context) {
     currentUserId() {
@@ -34,6 +42,7 @@ class _CreatePostState extends State<CreatePost> {
         child: Scaffold(
           key: viewModel.scaffoldKey,
           appBar: AppBar(
+            backgroundColor: Palette.mainBlueTheme,
             leading: IconButton(
               icon: Icon(Feather.x),
               onPressed: () {
@@ -41,24 +50,42 @@ class _CreatePostState extends State<CreatePost> {
                 Navigator.pop(context);
               },
             ),
-            title: Text('FlutterSocial'.toUpperCase()),
+            title: Text(
+              'Tạo bài viết',
+              style: MobileTextTheme().appBarStyle,
+            ),
             centerTitle: true,
             actions: [
               GestureDetector(
                 onTap: () async {
-                  await viewModel.uploadPosts(context);
-                  Navigator.pop(context);
-                  viewModel.resetPost();
+                  String input = description.text;
+
+                  if (cleanDescription(input)) {
+                    await viewModel.uploadPosts(context);
+                    viewModel.resetPost();
+                  } else {
+                    AwesomeDialog(
+                      context: context,
+                      dialogType: DialogType.WARNING,
+                      headerAnimationLoop: false,
+                      animType: AnimType.TOPSLIDE,
+                      showCloseIcon: true,
+                      closeIcon: const Icon(Icons.close_fullscreen_outlined),
+                      title: 'Cảnh báo!',
+                      desc:
+                          'Mô tả bạn nhập có chứa ký tự vi phạm tiêu chuẩn cộng đồng của chúng tôi',
+                      descTextStyle: AppTextTheme.oswaldTextStyle,
+                      btnCancelOnPress: () {},
+                      btnCancelText: 'Hủy bỏ',
+                    ).show();
+                  }
                 },
                 child: Padding(
-                  padding: const EdgeInsets.all(20.0),
+                  padding:
+                      const EdgeInsets.only(left: 20.0, right: 20.0, top: 20.0),
                   child: Text(
-                    'Post'.toUpperCase(),
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 15.0,
-                      color: Theme.of(context).accentColor,
-                    ),
+                    'Đăng bài',
+                    style: MobileTextTheme().appBarActionButton,
                   ),
                 ),
               )
@@ -80,105 +107,18 @@ class _CreatePostState extends State<CreatePost> {
                       ),
                       title: Text(
                         user?.username,
-                        style: TextStyle(fontWeight: FontWeight.bold),
+                        style: MobileTextTheme().currentUsernameTitle,
                       ),
                       subtitle: Text(
                         user?.email,
+                        style: MobileTextTheme().currentEmailTitle,
                       ),
                     );
                   }
                   return Container();
                 },
               ),
-              InkWell(
-                onTap: () => showImageChoices(context, viewModel),
-                child: Container(
-                  width: MediaQuery.of(context).size.width,
-                  height: MediaQuery.of(context).size.width - 30,
-                  decoration: BoxDecoration(
-                    color: Colors.grey[300],
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(5.0),
-                    ),
-                    border: Border.all(
-                      color: Theme.of(context).accentColor,
-                    ),
-                  ),
-                  child: viewModel.imgLink != null
-                      ? CustomImage(
-                          imageUrl: viewModel.imgLink,
-                          width: MediaQuery.of(context).size.width,
-                          height: MediaQuery.of(context).size.width - 30,
-                          fit: BoxFit.cover,
-                        )
-                      : viewModel.mediaUrl == null
-                          ? Center(
-                              child: Text(
-                                'Upload a Photo',
-                                style: TextStyle(
-                                  color: Theme.of(context).accentColor,
-                                ),
-                              ),
-                            )
-                          : Image.file(
-                              viewModel.mediaUrl,
-                              width: MediaQuery.of(context).size.width,
-                              height: MediaQuery.of(context).size.width - 30,
-                              fit: BoxFit.cover,
-                            ),
-                ),
-              ),
-              SizedBox(height: 20.0),
-              Text(
-                'Post Caption'.toUpperCase(),
-                style: TextStyle(
-                  fontSize: 15.0,
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-              TextFormField(
-                initialValue: viewModel.description,
-                decoration: InputDecoration(
-                  hintText: 'Eg. This is very beautiful place!',
-                  focusedBorder: UnderlineInputBorder(),
-                ),
-                maxLines: null,
-                onChanged: (val) => viewModel.setDescription(val),
-              ),
-              SizedBox(height: 20.0),
-              Text(
-                'Location'.toUpperCase(),
-                style: TextStyle(
-                  fontSize: 15.0,
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-              ListTile(
-                contentPadding: EdgeInsets.all(0.0),
-                title: Container(
-                  width: 250.0,
-                  child: TextFormField(
-                    controller: viewModel.locationTEC,
-                    decoration: InputDecoration(
-                      contentPadding: EdgeInsets.all(0.0),
-                      hintText: 'United States,Los Angeles!',
-                      focusedBorder: UnderlineInputBorder(),
-                    ),
-                    maxLines: null,
-                    onChanged: (val) => viewModel.setLocation(val),
-                  ),
-                ),
-                trailing: IconButton(
-                  tooltip: "Use your current location",
-                  icon: Icon(
-                    CupertinoIcons.map_pin_ellipse,
-                    size: 25.0,
-                  ),
-                  iconSize: 30.0,
-                  color: Theme.of(context).accentColor,
-                  onPressed: () => viewModel.getLocation(),
-                ),
-              ),
+              buildForm(context, viewModel),
             ],
           ),
         ),
@@ -186,6 +126,138 @@ class _CreatePostState extends State<CreatePost> {
     );
   }
 
+  buildForm(BuildContext context, PostsViewModel viewModel) {
+    return Form(
+      key: viewModel.formKey,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          InkWell(
+            onTap: () => showImageChoices(context, viewModel),
+            child: Container(
+              width: MediaQuery.of(context).size.width,
+              height: MediaQuery.of(context).size.width - 30,
+              decoration: BoxDecoration(
+                color: Palette.grey300,
+                borderRadius: BorderRadius.all(
+                  Radius.circular(5.0),
+                ),
+                border: Border.all(
+                  color: Palette.mainBlueTheme,
+                ),
+              ),
+              child: viewModel.imgLink != null
+                  ? CustomImage(
+                      imageUrl: viewModel.imgLink,
+                      width: MediaQuery.of(context).size.width,
+                      height: MediaQuery.of(context).size.width - 30,
+                      fit: BoxFit.cover,
+                    )
+                  : viewModel.mediaUrl == null
+                      ? Center(
+                          child: Text(
+                            'Nhấn vào đây để tải hình ảnh lên',
+                            style: MobileTextTheme().choosePictureRequired,
+                          ),
+                        )
+                      : Image.file(
+                          viewModel.mediaUrl,
+                          width: MediaQuery.of(context).size.width,
+                          height: MediaQuery.of(context).size.width - 30,
+                          fit: BoxFit.cover,
+                        ),
+            ),
+          ),
+          SizedBox(height: 20.0),
+          Row(
+            children: [
+              Text(
+                'Mô tả bài viết'.toUpperCase(),
+                style: MobileTextTheme().inputDescriptionAndLocationTitle,
+              ),
+              SizedBox(width: 5),
+              Text(
+                '*',
+                style: TextStyle(
+                  color: Palette.red,
+                ),
+              ),
+            ],
+          ),
+          TextFormField(
+            style: MobileTextTheme().inputDescriptionAndLocation,
+            controller: description,
+            decoration: InputDecoration(
+              hintText: 'Eg. Đây là một bức hình đẹp',
+              focusedBorder: UnderlineInputBorder(),
+            ),
+            validator: (value) {
+              if (value.trim().isEmpty) {
+                return ("Mời bạn nhập mô tả bài viết");
+              }
+              return null;
+            },
+            onSaved: (value) {
+              description.text = value;
+            },
+            maxLines: null,
+            onChanged: (val) => viewModel.setDescription(val),
+          ),
+          SizedBox(height: 20.0),
+          Text(
+            'Vị trí'.toUpperCase(),
+            style: MobileTextTheme().inputDescriptionAndLocationTitle,
+          ),
+          ListTile(
+            contentPadding: EdgeInsets.all(0.0),
+            title: Container(
+              width: 250.0,
+              child: TextFormField(
+                style: MobileTextTheme().inputDescriptionAndLocation,
+                controller: viewModel.locationTEC,
+                decoration: InputDecoration(
+                  contentPadding: EdgeInsets.all(0.0),
+                  hintText: 'Việt Nam, Hồ Chí Minh',
+                  focusedBorder: UnderlineInputBorder(),
+                ),
+                maxLines: null,
+                onChanged: (val) => viewModel.setLocation(val),
+              ),
+            ),
+            trailing: IconButton(
+              tooltip: "Sử dụng vị trí hiện tại của bạn",
+              icon: Icon(
+                CupertinoIcons.map_pin_ellipse,
+                size: 25.0,
+              ),
+              iconSize: 30.0,
+              color: Palette.mainBlueTheme,
+              onPressed: () => viewModel.getLocation(),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  /// Check bad word comment
+  bool cleanDescription(String descriptionInput) {
+    List<String> inputArray = descriptionInput.split(" ");
+    bool result = true;
+    for (final item in inputArray) {
+      for (final badWord in Validations.badWord) {
+        if (item.toLowerCase() == badWord) {
+          print(item.toLowerCase());
+          print(badWord);
+          print('Test');
+          result = false;
+        }
+      }
+    }
+    return result;
+  }
+
+  /// Show box select image choice
   showImageChoices(BuildContext context, PostsViewModel viewModel) {
     showModalBottomSheet(
       context: context,
@@ -202,16 +274,17 @@ class _CreatePostState extends State<CreatePost> {
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 10.0),
                 child: Text(
-                  'Select Image',
-                  style: TextStyle(
-                    fontWeight: FontWeight.bold,
-                  ),
+                  'Lựa chọn hình ảnh',
+                  style: MobileTextTheme().chooseImageTitle,
                 ),
               ),
               Divider(),
               ListTile(
                 leading: Icon(Feather.camera),
-                title: Text('Camera'),
+                title: Text(
+                  'Chọn từ máy ảnh',
+                  style: MobileTextTheme().selectCamera,
+                ),
                 onTap: () {
                   Navigator.pop(context);
                   viewModel.pickImage(camera: true);
@@ -219,7 +292,10 @@ class _CreatePostState extends State<CreatePost> {
               ),
               ListTile(
                 leading: Icon(Feather.image),
-                title: Text('Gallery'),
+                title: Text(
+                  'Chọn từ thư viện',
+                  style: MobileTextTheme().selectGallery,
+                ),
                 onTap: () {
                   Navigator.pop(context);
                   viewModel.pickImage();
@@ -229,6 +305,19 @@ class _CreatePostState extends State<CreatePost> {
           ),
         );
       },
+    );
+  }
+
+  postSuccessfully(context) async {
+    showInSnackBar('Không có kết nối mạng', context);
+  }
+
+  void showInSnackBar(String value, context) {
+    ScaffoldMessenger.of(context).removeCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(value),
+      ),
     );
   }
 }
