@@ -1,6 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import '/utils/validation.dart';
+import '/widgets/awesome_dialog.dart';
+import '/theme/palette.dart';
+import '/utils/app_text_theme.dart';
 import '/components/stream_comments_wrapper.dart';
 import '/models/comments.dart';
 import '/models/post.dart';
@@ -22,8 +26,15 @@ class _CommentsState extends State<Comments> {
   UserModel user;
 
   PostService services = PostService();
+
+  /// Get the current datetime
   final DateTime timestamp = DateTime.now();
+
+  /// Editting Controller
   TextEditingController commentsTEC = TextEditingController();
+
+  /// Check validation
+  GlobalKey<FormState> formKey = GlobalKey<FormState>();
 
   currentUserId() {
     return firebaseAuth.currentUser.uid;
@@ -42,73 +53,96 @@ class _CommentsState extends State<Comments> {
           ),
         ),
         centerTitle: true,
-        title: Text('Comments'),
+        backgroundColor: Palette.mainBlueTheme,
+        title: Text(
+          'Bình luận bài viết',
+          style: MobileTextTheme().appBarStyle,
+        ),
       ),
       body: Container(
         height: MediaQuery.of(context).size.height,
         child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             Flexible(
               child: ListView(
                 children: [
                   Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 10.0),
+                    padding: EdgeInsets.only(top: 20),
                     child: buildFullPost(),
                   ),
                   Divider(thickness: 1.5),
-                  Flexible(
-                    child: buildComments(),
-                  )
+                  buildComments(),
+                  commentWrapper(),
                 ],
               ),
             ),
-            Align(
-              alignment: Alignment.bottomCenter,
-              child: Container(
-                decoration: BoxDecoration(
-                  color: Theme.of(context).primaryColor,
-                ),
-                constraints: BoxConstraints(
-                  maxHeight: 190.0,
-                ),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    Flexible(
-                      child: ListTile(
-                        contentPadding: EdgeInsets.all(0),
-                        title: TextField(
-                          textCapitalization: TextCapitalization.sentences,
-                          controller: commentsTEC,
-                          style: TextStyle(
-                            fontSize: 15.0,
-                            color: Theme.of(context).textTheme.headline6.color,
+          ],
+        ),
+      ),
+    );
+  }
+
+  /// Check validation then upload comment
+  commentWrapper() {
+    return Form(
+      key: formKey,
+      child: Align(
+        alignment: Alignment.bottomCenter,
+        child: Container(
+          decoration: BoxDecoration(),
+          constraints: BoxConstraints(
+            maxHeight: 190.0,
+          ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              Flexible(
+                child: Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: ListTile(
+                    contentPadding: EdgeInsets.all(0),
+                    title: TextFormField(
+                      textCapitalization: TextCapitalization.sentences,
+                      controller: commentsTEC,
+                      style: TextStyle(
+                        fontSize: 15.0,
+                        color: Theme.of(context).textTheme.headline6.color,
+                      ),
+                      validator: (value) {
+                        if (value.trim().isEmpty) {
+                          return 'Mời bạn nhập nội dung bình luận';
+                        }
+                        return null;
+                      },
+                      decoration: InputDecoration(
+                        contentPadding: EdgeInsets.all(10.0),
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(5.0),
+                          borderSide: BorderSide(
+                            color: Theme.of(context).primaryColor,
                           ),
-                          decoration: InputDecoration(
-                            contentPadding: EdgeInsets.all(10.0),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(5.0),
-                              borderSide: BorderSide(
-                                color: Theme.of(context).primaryColor,
-                              ),
-                            ),
-                            enabledBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: Theme.of(context).primaryColor,
-                              ),
-                              borderRadius: BorderRadius.circular(5.0),
-                            ),
-                            hintText: "Write your comment...",
-                            hintStyle: TextStyle(
-                              fontSize: 15.0,
-                              color:
-                                  Theme.of(context).textTheme.headline6.color,
-                            ),
-                          ),
-                          maxLines: null,
                         ),
-                        trailing: GestureDetector(
-                          onTap: () async {
+                        enabledBorder: OutlineInputBorder(
+                          borderSide: BorderSide(
+                            color: Theme.of(context).primaryColor,
+                          ),
+                          borderRadius: BorderRadius.circular(5.0),
+                        ),
+                        hintText: "Viết bình luận của bạn...",
+                        hintStyle: TextStyle(
+                          fontSize: 15.0,
+                          color: Theme.of(context).textTheme.headline6.color,
+                        ),
+                      ),
+                      maxLines: null,
+                    ),
+                    trailing: GestureDetector(
+                      onTap: () async {
+                        String input = commentsTEC.text;
+
+                        if (cleanComment(input)) {
+                          if (formKey.currentState.validate()) {
                             await services.uploadComment(
                               currentUserId(),
                               commentsTEC.text,
@@ -116,26 +150,57 @@ class _CommentsState extends State<Comments> {
                               widget.post.ownerId,
                               widget.post.mediaUrl,
                             );
-                            commentsTEC.clear();
-                          },
-                          child: Padding(
-                            padding: const EdgeInsets.only(right: 10.0),
-                            child: Icon(
-                              Icons.send,
-                              color: Theme.of(context).accentColor,
-                            ),
-                          ),
+                          }
+                        } else {
+                          AwesomeDialog(
+                            context: context,
+                            dialogType: DialogType.WARNING,
+                            headerAnimationLoop: false,
+                            animType: AnimType.TOPSLIDE,
+                            showCloseIcon: true,
+                            closeIcon:
+                                const Icon(Icons.close_fullscreen_outlined),
+                            title: 'Cảnh báo!',
+                            desc:
+                                'Bình luận bạn nhập có chứa nội dung vi phạm tiêu chuẩn cộng đồng của chúng tôi',
+                            descTextStyle: AppTextTheme.oswaldTextStyle,
+                            btnOkOnPress: () {},
+                          ).show();
+                        }
+                        commentsTEC.clear();
+                      },
+                      child: Padding(
+                        padding: const EdgeInsets.only(right: 10.0),
+                        child: Icon(
+                          Icons.send,
                         ),
                       ),
                     ),
-                  ],
+                  ),
                 ),
               ),
-            ),
-          ],
+            ],
+          ),
         ),
       ),
     );
+  }
+
+  /// Check bad word comment
+  bool cleanComment(String commentInput) {
+    List<String> inputArray = commentInput.split(" ");
+    bool result = true;
+    for (final item in inputArray) {
+      for (final badWord in Validations.badWord) {
+        if (item.toLowerCase() == badWord) {
+          print(item.toLowerCase());
+          print(badWord);
+          print('Test');
+          result = false;
+        }
+      }
+    }
+    return result;
   }
 
   buildFullPost() {
@@ -145,53 +210,48 @@ class _CommentsState extends State<Comments> {
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         Container(
-          height: 250.0,
-          width: MediaQuery.of(context).size.width - 20.0,
+          height: MediaQuery.of(context).size.width,
+          width: MediaQuery.of(context).size.width,
           child: cachedNetworkImage(widget.post.mediaUrl),
         ),
         Padding(
           padding: const EdgeInsets.all(8.0),
-          child: Row(
+          child: Column(
+/*            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.start,*/
             children: [
-              Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                crossAxisAlignment: CrossAxisAlignment.start,
+              Text(
+                widget.post.description,
+                style: MobileTextTheme().postDescription,
+              ),
+              SizedBox(height: 4.0),
+              Row(
                 children: [
+                  StreamBuilder(
+                    stream: likesRef
+                        .where('postId', isEqualTo: widget.post.postId)
+                        .snapshots(),
+                    builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
+                      if (snapshot.hasData) {
+                        QuerySnapshot snap = snapshot.data;
+                        List<DocumentSnapshot> docs = snap.docs;
+                        return buildLikesCount(context, docs?.length ?? 0);
+                      } else {
+                        return buildLikesCount(context, 0);
+                      }
+                    },
+                  ),
+                  buildLikeButton(),
+                  Spacer(),
                   Text(
-                    widget.post.description,
-                    style: TextStyle(
-                      fontWeight: FontWeight.w800,
+                    timeago.format(
+                      widget.post.timestamp.toDate(),
                     ),
+                    style: TextStyle(),
                   ),
-                  SizedBox(height: 4.0),
-                  Row(
-                    children: [
-                      Text(
-                        timeago.format(widget.post.timestamp.toDate()),
-                        style: TextStyle(),
-                      ),
-                      SizedBox(width: 3.0),
-                      StreamBuilder(
-                        stream: likesRef
-                            .where('postId', isEqualTo: widget.post.postId)
-                            .snapshots(),
-                        builder:
-                            (context, AsyncSnapshot<QuerySnapshot> snapshot) {
-                          if (snapshot.hasData) {
-                            QuerySnapshot snap = snapshot.data;
-                            List<DocumentSnapshot> docs = snap.docs;
-                            return buildLikesCount(context, docs?.length ?? 0);
-                          } else {
-                            return buildLikesCount(context, 0);
-                          }
-                        },
-                      ),
-                    ],
-                  ),
+                  SizedBox(width: 3.0),
                 ],
               ),
-              Spacer(),
-              buildLikeButton(),
             ],
           ),
         ),
@@ -234,7 +294,7 @@ class _CommentsState extends State<Comments> {
               padding: const EdgeInsets.symmetric(horizontal: 20.0),
               child: Text(
                 comments.comment,
-                style: TextStyle(fontWeight: FontWeight.w400),
+                style: MobileTextTheme().postDescription,
               ),
             ),
             Divider()
@@ -306,15 +366,17 @@ class _CommentsState extends State<Comments> {
           .doc(widget.post.ownerId)
           .collection('notifications')
           .doc(widget.post.postId)
-          .set({
-        "type": "like",
-        "username": user.username,
-        "userId": currentUserId(),
-        "userDp": user.photoUrl,
-        "postId": widget.post.postId,
-        "mediaUrl": widget.post.mediaUrl,
-        "timestamp": timestamp,
-      });
+          .set(
+        {
+          "type": "like",
+          "username": user.username,
+          "userId": currentUserId(),
+          "userDp": user.photoUrl,
+          "postId": widget.post.postId,
+          "mediaUrl": widget.post.mediaUrl,
+          "timestamp": timestamp,
+        },
+      );
     }
   }
 
@@ -329,9 +391,14 @@ class _CommentsState extends State<Comments> {
           .collection('notifications')
           .doc(widget.post.postId)
           .get()
-          .then((doc) => {
-                if (doc.exists) {doc.reference.delete()}
-              });
+          .then(
+            (doc) => {
+              if (doc.exists)
+                {
+                  doc.reference.delete(),
+                }
+            },
+          );
     }
   }
 }
