@@ -1,12 +1,10 @@
 import 'dart:io';
-
 import 'package:firebase_storage/firebase_storage.dart';
 import '/utils/file_utils.dart';
 import '/utils/firebase.dart';
 
 abstract class Service {
-
-  //function to upload images to firebase storage and retrieve the url.
+  /// Function to upload images to firebase storage and retrieve the url.
   Future<String> uploadImage(Reference ref, File file) async {
     String ext = FileUtils.getFileExtension(file);
     Reference storageReference = ref.child("${uuid.v4()}.$ext");
