@@ -1,18 +1,16 @@
 import 'dart:io';
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import '/models/user.dart';
 import '/services/services.dart';
 import '/utils/firebase.dart';
 
 class UserService extends Service {
-  
   //get the authenticated uis
   String currentUid() {
     return firebaseAuth.currentUser.uid;
   }
 
-//tells when the user is online or not and updates the last seen for the messages
+/// Tells when the user is online or not and updates the last seen for the messages
   setUserStatus(bool isOnline) {
     var user = firebaseAuth.currentUser;
     if (user != null) {
@@ -22,10 +20,13 @@ class UserService extends Service {
     }
   }
 
-
-//updates user profile in the Edit Profile Screen
-  updateProfile(
-      {File image, String username, String bio, String country}) async {
+/// Updates user profile in the Edit Profile Screen
+  updateProfile({
+    File image,
+    String username,
+    String bio,
+    String country,
+  }) async {
     DocumentSnapshot doc = await usersRef.doc(currentUid()).get();
     var users = UserModel.fromJson(doc.data());
     users?.username = username;
