@@ -1,9 +1,13 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import '/widgets/awesome_dialog.dart';
+import '/theme/palette.dart';
 import '/screens/mainscreen.dart';
 import '/services/auth_service.dart';
 import '/utils/validation.dart';
+import '/utils/app_text_theme.dart';
 
 class LoginViewModel extends ChangeNotifier {
   GlobalKey<ScaffoldState> scaffoldKey = GlobalKey<ScaffoldState>();
@@ -21,25 +25,57 @@ class LoginViewModel extends ChangeNotifier {
     if (!form.validate()) {
       validate = true;
       notifyListeners();
-      showInSnackBar('Please fix the errors in red before submitting.',context);
+      AwesomeDialog(
+        autoHide: Duration(seconds: 3),
+        btnCancelColor: Palette.warningColor,
+        context: context,
+        dialogType: DialogType.ERROR,
+        headerAnimationLoop: false,
+        animType: AnimType.TOPSLIDE,
+        showCloseIcon: true,
+        closeIcon: const Icon(Icons.close_fullscreen_outlined),
+        title: 'Thông báo!',
+        desc: 'Vui lòng hoàn thành điền thông tin trước khi đăng nhập. ',
+        descTextStyle: AppTextTheme.oswaldTextStyle,
+        btnCancelOnPress: () {},
+        btnCancelText: 'Hủy bỏ',
+      ).show();
     } else {
       loading = true;
       notifyListeners();
       try {
-        bool success = await auth.loginUser(
-          email: email,
-          password: password,
-        );
+        bool success = await auth
+            .loginUser(
+              email: email,
+              password: password,
+            );
         print(success);
         if (success) {
-          Navigator.of(context).pushReplacement(
-              CupertinoPageRoute(builder: (_) => TabScreen()));
+          AwesomeDialog(
+            btnCancelColor: Palette.warningColor,
+            context: context,
+            dialogType: DialogType.SUCCES,
+            headerAnimationLoop: false,
+            animType: AnimType.TOPSLIDE,
+            showCloseIcon: true,
+            closeIcon: const Icon(Icons.close_fullscreen_outlined),
+            desc: 'Đăng nhập thành công. ',
+            descTextStyle: AppTextTheme.oswaldTextStyle,
+            btnOkOnPress: () {
+              Navigator.of(context).pushReplacement(
+                CupertinoPageRoute(
+                  builder: (context) => TabScreen(),
+                ),
+              );
+            },
+          ).show();
         }
       } catch (e) {
         loading = false;
         notifyListeners();
         print(e);
-        showInSnackBar('${auth.handleFirebaseAuthError(e.toString())}',context);
+        showErrorInSnackBar(
+            '${auth.handleFirebaseAuthError(e.toString())}', context);
       }
       loading = false;
       notifyListeners();
@@ -53,14 +89,23 @@ class LoginViewModel extends ChangeNotifier {
     form.save();
     print(Validations.validateEmail(email));
     if (Validations.validateEmail(email) != null) {
-      showInSnackBar('Please input a valid email to reset your password.',context);
+      showErrorInSnackBar(
+        'Hãy nhập Email hợp lệ để lấy lại mật khẩu',
+        context,
+      );
     } else {
       try {
         await auth.forgotPassword(email);
-        showInSnackBar('Please check your email for instructions '
-            'to reset your password', context);
+        showErrorInSnackBar(
+          'Vui lòng kiểm tra email của bạn để được hướng dẫn '
+          'đặt lại mật khẩu',
+          context,
+        );
       } catch (e) {
-        showInSnackBar('${e.toString()}', context);
+        showErrorInSnackBar(
+          '${e.toString()}',
+          context,
+        );
       }
     }
     loading = false;
@@ -77,8 +122,31 @@ class LoginViewModel extends ChangeNotifier {
     notifyListeners();
   }
 
-  void showInSnackBar(String value,context) {
+  /// This message will show when validation is failed
+  void showErrorInSnackBar(String value, context) {
     ScaffoldMessenger.of(context).removeCurrentSnackBar();
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(value)));
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(
+          value,
+          style: GoogleFonts.oswald(),
+        ),
+        backgroundColor: Palette.red,
+      ),
+    );
+  }
+
+  /// This message will show when validation is successful
+  void showActiveInSnackBar(String value, context) {
+    ScaffoldMessenger.of(context).removeCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(
+          value,
+          style: GoogleFonts.oswald(),
+        ),
+        backgroundColor: Palette.activeButton,
+      ),
+    );
   }
 }
