@@ -3,22 +3,35 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-//import 'package:flutter_icons/flutter_icons.dart';
+import 'package:http/http.dart';
+import '/pages/nav-items/feeds/widgets/edit_post_screen.dart';
+import '/theme/palette.dart';
+import '/utils/app_text_theme.dart';
 import '/components/custom_card.dart';
 import '/components/custom_image.dart';
 import '/models/post.dart';
 import '/models/user.dart';
-import '/pages/profile.dart';
+import '/pages/drawer-items/profile/pages/profile.dart';
 import '/screens/comment.dart';
 import '/screens/view_image.dart';
 import '/services/post_service.dart';
 import '/utils/firebase.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
-class UserPost extends StatelessWidget {
+class UserPost extends StatefulWidget {
   final PostModel post;
+  final postId;
 
-  UserPost({this.post});
+  UserPost({
+    this.post,
+    this.postId,
+  });
+
+  @override
+  State<UserPost> createState() => _UserPostState();
+}
+
+class _UserPostState extends State<UserPost> {
   final DateTime timestamp = DateTime.now();
 
   currentUserId() {
@@ -35,7 +48,7 @@ class UserPost extends StatelessWidget {
       child: OpenContainer(
         transitionType: ContainerTransitionType.fadeThrough,
         openBuilder: (BuildContext context, VoidCallback _) {
-          return ViewImage(post: post);
+          return ViewImage(post: widget.post);
         },
         closedElevation: 0.0,
         closedShape: const RoundedRectangleBorder(
@@ -50,16 +63,14 @@ class UserPost extends StatelessWidget {
             children: [
               Column(
                 children: [
+                  buildUser(context),
+                  SizedBox(height: 5.0),
                   ClipRRect(
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(10.0),
-                      topRight: Radius.circular(10.0),
-                    ),
                     child: CustomImage(
-                      imageUrl: post?.mediaUrl ?? '',
-                      height: 300.0,
+                      imageUrl: widget.post?.mediaUrl ?? '',
+                      height: MediaQuery.of(context).size.width,
+                      width: MediaQuery.of(context).size.width,
                       fit: BoxFit.cover,
-                      width: double.infinity,
                     ),
                   ),
                   Padding(
@@ -77,7 +88,8 @@ class UserPost extends StatelessWidget {
                                 onTap: () {
                                   Navigator.of(context).push(
                                     CupertinoPageRoute(
-                                      builder: (_) => Comments(post: post),
+                                      builder: (_) =>
+                                          Comments(post: widget.post),
                                     ),
                                   );
                                 },
@@ -86,6 +98,7 @@ class UserPost extends StatelessWidget {
                                   size: 25.0,
                                 ),
                               ),
+                              buildCurrentUserEditPost(post),
                             ],
                           ),
                         ),
@@ -96,7 +109,8 @@ class UserPost extends StatelessWidget {
                                 padding: const EdgeInsets.only(left: 5.0),
                                 child: StreamBuilder(
                                   stream: likesRef
-                                      .where('postId', isEqualTo: post.postId)
+                                      .where('postId',
+                                          isEqualTo: widget.post.postId, )
                                       .snapshots(),
                                   builder: (context,
                                       AsyncSnapshot<QuerySnapshot> snapshot) {
@@ -115,7 +129,7 @@ class UserPost extends StatelessWidget {
                             SizedBox(width: 5.0),
                             StreamBuilder(
                               stream: commentRef
-                                  .doc(post.postId)
+                                  .doc(widget.post.postId)
                                   .collection("comments")
                                   .snapshots(),
                               builder: (context,
@@ -133,17 +147,14 @@ class UserPost extends StatelessWidget {
                           ],
                         ),
                         Visibility(
-                          visible: post.description != null &&
-                              post.description.toString().isNotEmpty,
+                          visible: widget.post.description != null &&
+                              widget.post.description.toString().isNotEmpty,
                           child: Padding(
-                            padding: const EdgeInsets.only(left: 10.0, top: 3.0),
+                            padding:
+                                const EdgeInsets.only(left: 10.0, top: 3.0),
                             child: Text(
-                              '${post?.description ?? ""}',
-                              style: TextStyle(
-                                color:
-                                    Theme.of(context).textTheme.caption.color,
-                                fontSize: 15.0,
-                              ),
+                              '${widget.post?.description ?? ""}',
+                              style: MobileTextTheme().postDescription,
                               maxLines: 2,
                             ),
                           ),
@@ -151,8 +162,12 @@ class UserPost extends StatelessWidget {
                         SizedBox(height: 3.0),
                         Padding(
                           padding: const EdgeInsets.all(3.0),
-                          child: Text(timeago.format(post.timestamp.toDate()),
-                              style: TextStyle(fontSize: 10.0)),
+                          child: Text(
+                            timeago.format(
+                              widget.post.timestamp.toDate(),
+                            ),
+                            style: MobileTextTheme().timePost,
+                          ),
                         ),
                         // SizedBox(height: 5.0),
                       ],
@@ -160,7 +175,6 @@ class UserPost extends StatelessWidget {
                   )
                 ],
               ),
-              buildUser(context),
             ],
           );
         },
@@ -171,7 +185,7 @@ class UserPost extends StatelessWidget {
   buildLikeButton() {
     return StreamBuilder(
       stream: likesRef
-          .where('postId', isEqualTo: post.postId)
+          .where('postId', isEqualTo: widget.post.postId)
           .where('userId', isEqualTo: currentUserId())
           .snapshots(),
       builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
@@ -182,14 +196,14 @@ class UserPost extends StatelessWidget {
               if (docs.isEmpty) {
                 likesRef.add({
                   'userId': currentUserId(),
-                  'postId': post.postId,
+                  'postId': widget.post.postId,
                   'dateCreated': Timestamp.now(),
                 });
                 addLikesToNotification();
               } else {
                 likesRef.doc(docs[0].id).delete();
                 services.removeLikeFromNotification(
-                    post.ownerId, post.postId, currentUserId());
+                    widget.post.ownerId, widget.post.postId, currentUserId());
               }
             },
             icon: docs.isEmpty
@@ -198,7 +212,7 @@ class UserPost extends StatelessWidget {
                   )
                 : Icon(
                     CupertinoIcons.heart_fill,
-                    color: Colors.red,
+                    color: Palette.red,
                   ),
           );
         }
@@ -208,13 +222,19 @@ class UserPost extends StatelessWidget {
   }
 
   addLikesToNotification() async {
-    bool isNotMe = currentUserId() != post.ownerId;
+    bool isNotMe = currentUserId() != widget.post.ownerId;
 
     if (isNotMe) {
       DocumentSnapshot doc = await usersRef.doc(currentUserId()).get();
       user = UserModel.fromJson(doc.data());
-      services.addLikesToNotification("like", user.username, currentUserId(),
-          post.postId, post.mediaUrl, post.ownerId, user.photoUrl);
+      services.addLikesToNotification(
+          "like",
+          user.username,
+          currentUserId(),
+          widget.post.postId,
+          widget.post.mediaUrl,
+          widget.post.ownerId,
+          user.photoUrl);
     }
   }
 
@@ -223,10 +243,7 @@ class UserPost extends StatelessWidget {
       padding: const EdgeInsets.only(left: 7.0),
       child: Text(
         '$count likes',
-        style: TextStyle(
-          fontWeight: FontWeight.bold,
-          fontSize: 10.0,
-        ),
+        style: MobileTextTheme().likesCount,
       ),
     );
   }
@@ -236,74 +253,92 @@ class UserPost extends StatelessWidget {
       padding: const EdgeInsets.only(top: 0.5),
       child: Text(
         '-   $count comments',
-        style: TextStyle(fontSize: 8.5, fontWeight: FontWeight.bold),
+        style: MobileTextTheme().commentCount,
+      ),
+    );
+  }
+
+  /// Edit post button
+  buildCurrentUserEditPost(post) {
+    /// Check if this post is your current authenticated
+    bool isMe = currentUserId() != widget.post.ownerId;
+    return Visibility(
+      visible: !isMe,
+      child: IconButton(
+        onPressed: () {
+          Navigator.of(context).push(
+            CupertinoPageRoute(
+              builder: (_) => EditPostScreen(
+                documentId: widget.post.postId,
+                currentUserID: widget.post.ownerId,
+                currentImageUrl: widget.post.mediaUrl,
+                currentDescription: widget.post.description,
+                currentLocation: widget.post.location,
+              ),
+            ),
+          );
+        },
+        icon: Icon(
+          Icons.edit_note,
+          color: Theme.of(context).iconTheme.color,
+        ),
       ),
     );
   }
 
   buildUser(BuildContext context) {
-    bool isMe = currentUserId() == post.ownerId;
     return StreamBuilder(
-      stream: usersRef.doc(post.ownerId).snapshots(),
+      stream: usersRef.doc(widget.post.ownerId).snapshots(),
       builder: (context, snapshot) {
         if (snapshot.hasData) {
           DocumentSnapshot snap = snapshot.data;
           UserModel user = UserModel.fromJson(snap.data());
-          return Visibility(
-            visible: !isMe,
-            child: Align(
-              alignment: Alignment.topCenter,
-              child: Container(
-                height: 40.0,
-                decoration: BoxDecoration(
-                  color: Colors.white60,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(10.0),
-                    topRight: Radius.circular(10.0),
-                  ),
+          return Align(
+            alignment: Alignment.topCenter,
+            child: Container(
+              height: 80.0,
+              decoration: BoxDecoration(
+                color: Palette.white60,
+                borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(10.0),
+                  topRight: Radius.circular(10.0),
                 ),
-                child: GestureDetector(
-                  onTap: () => showProfile(context, profileId: user?.id),
-                  child: Padding(
-                    padding: const EdgeInsets.only(left: 10.0),
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: [
-                        user.photoUrl.isNotEmpty
-                            ? CircleAvatar(
-                                radius: 14.0,
-                                backgroundColor: Color(0xff4D4D4D),
-                                backgroundImage: CachedNetworkImageProvider(
-                                    user?.photoUrl ?? ""),
-                              )
-                            : CircleAvatar(
-                                radius: 14.0,
-                                backgroundColor: Color(0xff4D4D4D),
-                              ),
-                        SizedBox(width: 5.0),
-                        Column(
-                          mainAxisSize: MainAxisSize.min,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              '${post?.username ?? ""}',
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                color: Color(0xff4D4D4D),
-                              ),
-                              overflow: TextOverflow.ellipsis,
+              ),
+              child: GestureDetector(
+                onTap: () => showProfile(context, profileId: user?.id),
+                child: Padding(
+                  padding: const EdgeInsets.only(left: 10.0),
+                  child: Row(
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      user.photoUrl.isNotEmpty
+                          ? CircleAvatar(
+                              radius: 14.0,
+                              backgroundColor: Palette.greyBlack,
+                              backgroundImage: CachedNetworkImageProvider(
+                                  user?.photoUrl ?? ""),
+                            )
+                          : CircleAvatar(
+                              radius: 14.0,
+                              backgroundColor: Palette.greyBlack,
                             ),
-                            Text(
-                              '${post?.location ?? 'Wooble'}',
-                              style: TextStyle(
-                                fontSize: 10.0,
-                                color: Color(0xff4D4D4D),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
+                      SizedBox(width: 5.0),
+                      Column(
+                        mainAxisSize: MainAxisSize.min,
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Text(
+                            '${widget.post?.username ?? ""}',
+                            style: MobileTextTheme().postUserNameTitle,
+                            overflow: TextOverflow.ellipsis,
+                          ),
+                          Text(
+                            '${widget.post?.location ?? 'Không có vị trí'}',
+                            style: MobileTextTheme().postLocationTitle,
+                          ),
+                        ],
+                      ),
+                    ],
                   ),
                 ),
               ),
@@ -316,7 +351,10 @@ class UserPost extends StatelessWidget {
     );
   }
 
-  showProfile(BuildContext context, {String profileId}) {
+  showProfile(
+    BuildContext context, {
+    String profileId,
+  }) {
     Navigator.push(
       context,
       CupertinoPageRoute(
